backend/requirements.txt
crewai>=0.130.0
python-dotenv>=1.0.1
openai>=1.0.0
litellm>=1.0.0
apscheduler>=3.10.4
fastapi>=0.116.0
uvicorn>=0.30.0
